changed integer types

+++ src/constant.h
@@ -11,8 +11,8 @@
 
 #include "headers.h"
 
-constexpr int32_t MAX_UCHAR = std::numeric_limits<uint8_t>::max();
-constexpr int32_t MAX_SHORT = std::numeric_limits<short>::max(); // maximum short/long signed ints
+constexpr uint8_t MAX_UCHAR = std::numeric_limits<uint8_t>::max();
+constexpr int16_t MAX_SHORT = std::numeric_limits<short>::max(); // maximum short/long signed ints
 constexpr int32_t MAX_LONG = std::numeric_limits<int32_t>::max();
 
 // Note to the Wizard:
